requirements.txt
boto3==1.4.4
celery[redis]==4.1.0
Django==1.11.3
django-admin-interface==0.5.1
django-annoying==0.10.3
django-celery-email==2.0.0
django-cors-headers==2.1.0
django-extensions==1.8.1
django-filter==1.0.4
django-mail-templated==2.6.5
-e git+https://github.com/PhilipGarnero/django-multiselectfield.git@2490e1c18fdcf127e214d6c6c70aab5800d9d5c3#egg=django_multiselectfield
django-rest-framework-social-oauth2==1.0.8
django-smuggler==0.8.0
django-storages==1.6.3
djangorestframework==3.6.3
gunicorn==19.7.1
ipython[notebook]==5.4.1
Markdown==2.6.8
MySQL-python==1.2.5
pillow==4.2.1
pycrypto==2.6.1
python-memcached==1.58
requests==2.18.1
unidecode==0.4.21
